Change output type vectorizer (#1)
* change setup

* change default output type of countvectorizer to bet in32

* change dtype to numerical encoder as well + tests

* add output type test on NumImputer

+++ aikit/transformers/categories.py
@@ -261,7 +261,7 @@ class _NumericalEncoder(BaseEstimator, TransformerMixin):
         ]
 
         if self.encoding_type == "num":
-            result = pd.DataFrame(all_result_series).T
+            result = pd.DataFrame(all_result_series, dtype="int32").T
 
             return result
